lets_part2

+++ ft_strlcpy.c
@@ -6,13 +6,15 @@
 /*   By: stakaki <stakaki@student.42tokyo.j>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/09 01:14:09 by stakaki           #+#    #+#             */
-/*   Updated: 2021/04/09 10:55:07 by stakaki          ###   ########.fr       */
+/*   Updated: 2021/04/12 13:44:23 by stakaki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_strlcpy(char *dest, char *src, unsigned int size)
+#include "libft.h"
+
+unsigned int	ft_strlcpy(char *dest, char *src, size_t size)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	if (size != 0)
@@ -22,7 +24,7 @@ unsigned int	ft_strlcpy(char *dest, char *src, unsigned int size)
 			dest[i] = src[i];
 			i++;
 		}
-	dest[i] = '\0';
+		dest[i] = '\0';
 	}
 	while (src[i] != '\0')
 		i++;
